Better refactoring for GraphConstruct operation
Before the "resources" endpoint parameter was used in multiple
operations but with different functionalities. For a GET operation it
was used to point to the resources containing the triples to load into
the newly created graph. For a CONSTRUCT operation it was used to
point to the SPARQL queries to execute on the graph.
This behaviour caused problems when the CONSTRUCT operation was run
with the newGraph flag/boolean. The user supplied a resource parameter
intended to hold the queries to run in the CONSTRUCT operation but it
was also used in the GET operation. This would fail because the
resources would point to queries and not to RDF triples.

The implemented solution is that of adding a QueryUrls parameter which
is only used to hold a reference to the queries. The resources
parameter is as before used to hold a reference to the file/s (or
urls) containing the triples.