Fix ContextAwareRepository bug
For some reason adding trying to add all the statements contained in a
rdf4j Model to a ContextAwareReposity and thus ContextAwareConnection
does not work and result in a stack overflow error.
The implemented fix is to iterate over the statements in the model and
add them to the repository one by one.