fix: usage of user supplied TempFunctions
The supplied file MUST be called "CustomFunctions.java".  This is
because in java a class and the file that contains that class must
have the same name. Because a temp file that is a copy of the original
"*.java" file is created, i have chosen to call this temp file
"CustomFunctions.java".