refactor ChimeraResourceBean, no refs to other beans
Removed references to other beans that were used to handle the
authentication detail for remote resources. Now the idea is that the
ChimeraResourceBean has attributes for all possible authentication
parameters. The user is able to specify the desired ones. Then the
Bean is "specialized" to one of the authentication types that are
currently supported, i.e. the 'TypeAuthConfig" interface.