fix: single namedGraph set as insert context for ContextAwareRepo
A ContextAwareRepository can have a single context (namedGraph) either
because the user passed one namedGraph or because the namedGraph was
generated by us using the Exchange id. To handle this latter case we
have to set the insert context so that triples written in a file
without specifying a graph can be inserted in the randomly generated
namedGraph. This is because the user cannot possibly know in advance
the name of this randomly generated graph.