fix: creation of mapt singleton TemplateExecutor
Fixes the creation logic for the TemplateExecutor used to run MTL
mappings. Previously, due to faulty logic in the the if statement the
second time that the same mapping-template component object would be
used the template process would not be called, leading to incorrect
behaviour where the input exchange body would be returned as is by the
mapping-template component.